Example: Try to show a filename for data extents
Here's an example program that uses the logical_to_ino ioctl and the
ino_lookup ioctl to try show a filename for every inode that is reported
as using data from an extent in a block group.

The big limitation when doing this, is that logical_to_ino takes an
exact vaddr and only returns which inode is using that actual byte from
the extent.

If an inode is found by logical_to_ino, ino_lookup is called to get the
first filename that is known for this inode, relative to the top of the
subvolume the file is placed in.

-# examples/show_block_group_data_extent_filenames.py 289406976 /mnt/i320/fstree/
block group vaddr 289406976 transid 60 length 1073741824 flags DATA used 4530176 used_pct 0
extent vaddr 336048128 length 4186112 refs 1 gen 57 flags DATA
extent vaddr 341217280 length 16384 refs 1 gen 60 flags DATA
    root 5 inode 266 offset 0 path 'vmlinuz-4.9.0-1-amd64'
extent vaddr 341233664 length 65536 refs 1 gen 60 flags DATA
    root 1 inode 256 offset 0
extent vaddr 341299200 length 262144 refs 1 gen 60 flags DATA
    root 1 inode 259 offset 0

For the first extent at 336048128 logical_to_ino does not report any
inode using it. In this example, the first 4186112 byte extent is
actually 'vmlinuz-4.9.0-1-amd64', which got the first 16kiB overwritten
with garbage (the new 16kiB extent listed). :(

It would of course be possible to brute force this by incrementing the
vaddr with 4k until a filename is found, but doing so is left as an
exercise for the reader. ;-]