Add v2 search ioctl
The v2 search ioctl has been around since kernel 3.16. It can handle
bigger buffers than 4096 bytes, which we need for retrieving metadata
objects that are bigger than that. Or, it can be used to just get more
data back in one go when doing a big tree walk.

An example is an inode_ref object which can hold many names if a file
gets hardlinked a lot. When hardlinking a file more than a few hundred
times, the search v1 ioctl would return the object header, but with size
0 for data, since it doesn't fit.